refactor: rename timestamp module to time
