+++ gateway-common/src/lib.rs
@@ -2,5 +2,5 @@ pub mod blocklist;
 pub mod caching;
 pub mod http_ext;
 pub mod ptr;
-pub mod timestamp;
+pub mod time;
 pub mod ttl_hash_map;
